+++ src/index.js
@@ -3,5 +3,5 @@ export * as fees from './fees';
 export * as pmc from './pmc';
 export * as taxes from './taxes';
 export * as units from './units';
-export * as images from './images';
+export * as media from './media';
 export * as availability from './availability';
